Revert to previous logic in actions

+++ src/main/java/teammates/ui/webapi/DeleteAccountRequestAction.java
@@ -1,7 +1,7 @@
 package teammates.ui.webapi;
 
-import teammates.common.exception.InvalidOperationException;
 import teammates.common.util.Const;
+import teammates.storage.sqlentity.AccountRequest;
 
 /**
  * Deletes an existing account request.
@@ -9,15 +9,19 @@ import teammates.common.util.Const;
 class DeleteAccountRequestAction extends AdminOnlyAction {
 
     @Override
-    public JsonResult execute() throws teammates.ui.webapi.InvalidOperationException {
+    public JsonResult execute() throws InvalidOperationException {
         String email = getNonNullRequestParamValue(Const.ParamsNames.INSTRUCTOR_EMAIL);
         String institute = getNonNullRequestParamValue(Const.ParamsNames.INSTRUCTOR_INSTITUTION);
 
-        try {
-            sqlLogic.deleteAccountRequest(email, institute);
-        } catch (InvalidOperationException ioe) {
-            throw new teammates.ui.webapi.InvalidOperationException(ioe);
+        AccountRequest toDelete = sqlLogic.getAccountRequest(email, institute);
+
+        if (toDelete != null && toDelete.getRegisteredAt() != null) {
+            // instructor is already registered and cannot be deleted
+            throw new InvalidOperationException("Account request of a registered instructor cannot be deleted.");
         }
+
+        sqlLogic.deleteAccountRequest(email, institute);
+
         return new JsonResult("Account request successfully deleted.");
     }
